Add files via upload
Water-Quality-Prediction-by-using-Machine-Learning.

The main purpose of our proposed method is used to predict the quality of water by using Machine Learning algorithm. Here, we use the Five module to implement our article are Datasets collection, Data Preprocessing, Data Visualization, Model Implementation, Classification and Prediction. First module is the dataset collection, the datasets are generated with the help of the hardware components like sensor namely Temperature sensor is used to find the Temperature value in water, Ph sensor is used to find the Ph value in water and light intensity sensor is used to find the Turbidity value in water. Second module is data preprocessing method are used to process the data into the proper format of the system. Third module is the data visualization module to plot our data like pie chart, countplot, distplot, kdeplot, Data correlation and Pandas Profile report with the help of Exploratory data analysis concept. Fourth module is the model implementation, here two types of machine learning algorithm like Random Forest and Gaussian Naive Bayes are used to train and fit the model of our system. Fifth module is the classification and prediction model is used to classify and predict the data based on water quality or not. Finally, the performance metrics are determined accuracy score and classification report like precision, recall, f1 score and confusion matrix of both the algorithm as well as shown the comparsion graph of both algorithm of the system.